Add test for help text handling
